requirements.txt
langchain-core==0.2.11
langchain-openai==0.1.14
langchain-upstage==0.1.7
python-dotenv==1.0.1
